make tenant view refresh interval configurable
