+++ src/Beckett/Dashboard/Postgres/Tenants/ITenantMaterializedViewManager.cs
@@ -1,0 +1,6 @@
+namespace Beckett.Dashboard.Postgres.Tenants;
+
+public interface ITenantMaterializedViewManager
+{
+    Task Refresh(CancellationToken cancellationToken);
+}
