+++ src/Beckett/Dashboard/Postgres/Tenants/RefreshTenantMaterializedView.cs
@@ -1,14 +1,11 @@
 using Beckett.Database;
-using Beckett.Database.Queries;
 using Microsoft.Extensions.Hosting;
 using Microsoft.Extensions.Logging;
-using Npgsql;
 
 namespace Beckett.Dashboard.Postgres.Tenants;
 
 public class RefreshTenantMaterializedView(
-    IPostgresDataSource dataSource,
-    IPostgresDatabase database,
+    ITenantMaterializedViewManager tenantMaterializedViewManager,
     PostgresOptions options,
     ILogger<RefreshTenantMaterializedView> logger
 ) : BackgroundService
@@ -23,26 +20,9 @@ public class RefreshTenantMaterializedView(
             {
                 stoppingToken.ThrowIfCancellationRequested();
 
-                const string key = "Beckett:RefreshTenantMaterializedView";
+                await tenantMaterializedViewManager.Refresh(stoppingToken);
 
-                await using var connection = dataSource.CreateConnection();
-
-                await connection.OpenAsync(stoppingToken);
-
-                var locked = await database.Execute(new TryAdvisoryLock(key, options), connection, stoppingToken);
-
-                if (locked)
-                {
-                    logger.LogTrace("Starting tenant materialized view refresh");
-
-                    await database.Execute(new Query(options), connection, stoppingToken);
-
-                    await database.Execute(new AdvisoryUnlock(key, options), connection, stoppingToken);
-
-                    logger.LogTrace("Tenant materialized view refresh completed");
-                }
-
-                await Task.Delay(TimeSpan.FromMinutes(5), stoppingToken);
+                await Task.Delay(options.TenantRefreshInterval, stoppingToken);
             }
             catch (OperationCanceledException e) when (e.CancellationToken == stoppingToken)
             {
@@ -52,18 +32,6 @@ public class RefreshTenantMaterializedView(
             {
                 logger.LogError(e, "Error refreshing tenant materialized view");
             }
-        }
-    }
-
-    private class Query(PostgresOptions options) : IPostgresDatabaseQuery<int>
-    {
-        public Task<int> Execute(NpgsqlCommand command, CancellationToken cancellationToken)
-        {
-            command.CommandText = $"REFRESH MATERIALIZED VIEW CONCURRENTLY {options.Schema}.tenants;";
-
-            command.CommandTimeout = 120;
-
-            return command.ExecuteNonQueryAsync(cancellationToken);
         }
     }
 }
